Refactor messageEventContent APIs
